Add first semifunctioning version of PIPS-SBB.
Add first draft of PIPS-SBB: PIPS simple branch & bound. The branch
and bound code compiles, but the current driver executable does not
use it. In this sense, the code committed "works" -- the LP
relaxation of the MIP will solve successfully in the driver --
but the branch and bound code must still be shored up.