PIPS-SBB: Comment out erroneous B&B tree LB update
The branch-and-bound (B&B) tree lower bound (LB) update is commented
out because the code carrying out the update is wrong. The mixed
integer program (MIP) LB is the minimum of the B&B subtree LBs, which
are each linear program (LP) LBs. I don't currently track this
minimum; I should, if I want to stop the solver due to an optimality
gap tolerance.